refactored equipment files
