+++ app/src/main/java/com/example/sportshub/equipment/EquipmentViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.sportshub.dashboard
+package com.example.sportshub.equipment
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class EquipmentViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is dashboard Fragment"
