+++ app/src/main/res/layout/fragment_equipment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".dashboard.DashboardFragment">
+    tools:context=".equipment.EquipmentFragment">
 
     <TextView
         android:id="@+id/text_dashboard"
